refactor(EG-941): de-duplicate file upload step components across seqera and omics
